refactor: move tx output parsing to separate file
