+++ packages/cardano-metamask-snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/nufi-official/nufi-snap"
   },
   "source": {
-    "shasum": "5+D9NhyMZgehSwH1HBvFdq6OeLZu/bG5QB2becofD1E=",
+    "shasum": "XRSmKaCNa0xSM5KyVCm4P08gzhqoZfym/OYE+PsZLus=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
@@ -26,11 +26,7 @@
     "snap_getEntropy": {},
     "snap_getBip32Entropy": [
       {
-        "path": [
-          "m",
-          "1852'",
-          "1815'"
-        ],
+        "path": ["m", "1852'", "1815'"],
         "curve": "ed25519Bip32"
       }
     ]
