+++ packages/cardano-metamask-snap/src/api/sdk/transaction.ts
@@ -1,4 +1,4 @@
-import { Cardano, Serialization, type TxCBOR } from '@cardano-sdk/core';
+import { Serialization, type TxCBOR } from '@cardano-sdk/core';
 import { blake2b } from '@cardano-sdk/crypto';
 import { assert } from '@metamask/snaps-sdk';
 
@@ -11,7 +11,8 @@ import type {
 import { type CardanoDerivationPath } from '../derivationPath';
 import { parseCertificates } from './certificate';
 import type { TokenList } from './tokenList';
-import { applyDecimals, hexToBytes, lovelaceToAda } from './utils';
+import { hexToBytes, lovelaceToAda } from './utils';
+import { parseOutputs } from './output';
 
 /**
  * Calculates the hash of a transaction body.
@@ -79,49 +80,27 @@ export const parseTransaction = ({
   tokenList,
   ownCredentials,
 }: ParseTransactionParams): ParsedTransaction => {
-  const parsedTransaction = Serialization.Transaction.fromCbor(
+  const parsedTransactionBody = Serialization.Transaction.fromCbor(
     txCborHex as TxCBOR,
   ).body();
 
-  assertValidNetworkId(networkId, parsedTransaction);
+  assertValidNetworkId(networkId, parsedTransactionBody);
 
-  const outputs = parsedTransaction.outputs().map((output) => {
-    const address = output.address().toBech32();
-    return {
-      isChange: ownAddresses.some(
-        (ownAddress) => ownAddress.address === address,
-      ),
-      address,
-      coin: lovelaceToAda(output.amount().coin().toString()),
-      tokenBundle: Array.from(
-        output.amount().multiasset()?.entries() ?? [],
-      ).map(([assetId, value]) => {
-        const policyId = Cardano.AssetId.getPolicyId(assetId);
-        const assetName = Cardano.AssetId.getAssetName(assetId);
-        const fingerprint = Cardano.AssetFingerprint.fromParts(
-          policyId,
-          assetName,
-        ).toString();
-        const tokenMetadata = tokenList[fingerprint];
-        return {
-          fingerprint,
-          amount: applyDecimals(value.toString(), tokenMetadata?.decimals ?? 0),
-          name: tokenMetadata?.name,
-          ticker: tokenMetadata?.ticker,
-        };
-      }),
-    };
-  });
+  const outputs = parseOutputs(
+    parsedTransactionBody.outputs(),
+    ownAddresses,
+    tokenList,
+  );
 
   const certificates = parseCertificates(
-    parsedTransaction.certs(),
+    parsedTransactionBody.certs(),
     ownCredentials,
   );
 
-  const ttl = parsedTransaction.ttl()?.toString();
-  const validityIntervalStart = parsedTransaction
+  const ttl = parsedTransactionBody.ttl()?.toString();
+  const validityIntervalStart = parsedTransactionBody
     .validityStartInterval()
     ?.toString();
-  const fee = lovelaceToAda(parsedTransaction.fee().toString());
+  const fee = lovelaceToAda(parsedTransactionBody.fee().toString());
   return { outputs, fee, certificates, ttl, validityIntervalStart };
 };
